Refactor the codes and authentication system

+++ todo/routes/api.py
@@ -1,79 +1,219 @@
 from flask import Blueprint, request, jsonify, flash
+from flask_jwt_extended import jwt_required, current_user
 import json
 
-from ..extensions import db
-from ..models import *
+from ..extensions import db, jwt_manager
+from ..models import Users, Projects, Todos
 
 api_bp = Blueprint("api", __name__, url_prefix="/api")
 
 
-@api_bp.route("/users", methods=["GET", "POST"], strict_slashes=False)
+@jwt_manager.user_lookup_loader
+def user_loader(jwt_header, jwt_payload):
+    """
+    A function to get a logged in user data so it can be accessed from the 'current_user' object
+    """
+    identity = jwt_payload["sub"]
+    user = db.session.execute(db.select(Users).filter_by(user_id=identity)).scalar_one()
+    return user
+
+
+##### USERS ENDPOINTS #####
+
+
+@api_bp.route("/users", methods=["GET"], strict_slashes=False)
+@jwt_required()
 def get_users():
-    users = db.session.execute(db.select(Users).order_by(Users.user_id)).scalars()
+    if current_user.role.upper() != "ADMIN":
+        return jsonify({"success": False, "message": "Unauthorized action"}), 403
+
+    users = db.session.execute(db.select(Users)).scalars()
     data = [user.serialize() for user in users]
 
-    return jsonify(data), 200
+    return jsonify({"success": True, "data": data}), 200
 
 
 @api_bp.route("/users/<int:user_id>", methods=["GET", "PUT"], strict_slashes=False)
-def get_user(user_id):
+@jwt_required()
+def access_user(user_id):
+    if current_user.user_id != user_id and current_user.role.upper() != "ADMIN":
+        return jsonify({"success": False, "message": "Unauthorized action"}), 403
+
     user = db.session.execute(db.select(Users).filter_by(user_id=user_id)).scalar_one()
     data = user.serialize()
 
     if request.method == "PUT":
-        updated_data = json.loads(request.get_data(as_text=True))
-        user.name = updated_data["name"]
-        user.role = updated_data["role"]
-        user.bio = updated_data["bio"]
+        try:
+            # updated_data = json.loads(request.get_data(as_text=True))
+            updated_data = request.get_json()
+            user.name = updated_data["name"]
+            user.role = updated_data["role"]
+            user.bio = updated_data["bio"]
 
-        db.session.commit()
-        updated_data = user.serialize()
+            db.session.commit()
+        except:
+            db.session.rollback()
 
-        flash("Your profile has been updated", category="success")
+            flash("Failed to add the profile", category="error")
 
-        return jsonify(updated_data), 201
+            return (
+                jsonify({"success": False, "message": "Failed to update the profile"}),
+                500,
+            )
+        else:
+            updated_data = user.serialize()
 
-    return jsonify(data)
+            flash("Your profile has been updated", category="success")
+
+            return jsonify({"success": True, "data": updated_data}), 201
+
+    return jsonify({"success": True, "data": data}), 200
 
 
-@api_bp.route("/users/<int:user_id>/projects", methods=["GET"], strict_slashes=False)
-def get_projects(user_id):
+##### PROJECTS ENDPOINTS #####
+
+
+@api_bp.route(
+    "/users/<int:user_id>/projects", methods=["GET", "POST"], strict_slashes=False
+)
+@jwt_required()
+def access_projects(user_id):
+    if current_user.user_id != user_id and current_user.role.upper() != "ADMIN":
+        return jsonify({"success": False, "message": "Unauthorized action"}), 403
+
+    if request.method == "POST":
+        new_data = request.get_json()
+        project = Projects(
+            title=new_data["title"],
+            description=new_data["description"],
+            user_id=user_id,
+        )
+
+        try:
+            db.session.add(project)
+            db.session.commit()
+        except:
+            db.session.rollback()
+
+            flash("Failed to add the project", category="error")
+
+            return (
+                jsonify({"success": False, "message": "Failed to add the project"}),
+                500,
+            )
+        else:
+            flash(f"Your new project has been added to the list!", category="success")
+
+            return jsonify({"success": True, "data": project.serialize()}), 201
+
     projects = db.session.execute(
         db.select(Projects).filter_by(user_id=user_id).order_by(Projects.project_id)
     ).scalars()
     data = [project.serialize() for project in projects]
 
-    return jsonify(data)
+    return jsonify({"success": True, "data": data}), 200
 
 
 @api_bp.route(
-    "/projects/<int:project_id>",
-    methods=["GET", "POST", "PUT", "DELETE"],
+    "/users/<int:user_id>/projects/<int:project_id>",
+    methods=["GET", "PUT", "DELETE"],
     strict_slashes=False,
 )
-def get_project(project_id):
+@jwt_required()
+def access_project(user_id, project_id):
+    if current_user.user_id != user_id and current_user.role.upper() != "ADMIN":
+        return jsonify({"success": False, "message": "Unauthorized action"}), 403
+
     project = db.session.execute(
         db.select(Projects).filter_by(project_id=project_id)
     ).scalar_one()
     data = project.serialize()
 
-    if request.method == "DELETE":
-        db.session.delete(project)
-        db.session.commit()
+    if request.method == "PUT":
+        try:
+            updated_data = request.get_json()
+            project.title = updated_data["title"]
+            project.description = updated_data["description"]
 
-        flash(f"Your project has been deleted!", category="error")
+            db.session.commit()
+        except:
+            db.session.rollback()
 
-        return jsonify(data), 201
+            flash("Failed to update the project", category="error")
 
-    return jsonify(data), 200
+            return (
+                jsonify({"success": False, "message": "Failed to update the project"}),
+                500,
+            )
+        else:
+            updated_data = project.serialize()
+
+            flash(f"Your project has been updated!", category="success")
+
+            return jsonify({"success": True, "data": updated_data}), 201
+
+    elif request.method == "DELETE":
+        try:
+            db.session.delete(project)
+            db.session.commit()
+        except:
+            db.session.rollback()
+
+            flash("Failed to delete the project", category="error")
+
+            return (
+                jsonify({"success": False, "message": "Failed to delete the project"}),
+                500,
+            )
+        else:
+            flash(f"Your project has been deleted!", category="error")
+
+            return (
+                jsonify({"success": True, "message": "Your project has been deleted!"}),
+                201,
+            )
+
+    return jsonify({"success": True, "data": data}), 200
+
+
+##### TODOS ENDPOINTS #####
 
 
 @api_bp.route(
     "/users/<int:user_id>/todos",
-    methods=["GET"],
+    methods=["GET", "POST"],
     strict_slashes=False,
 )
-def get_all_todos(user_id):
+@jwt_required()
+def access_all_todos(user_id):
+    if current_user.user_id != user_id and current_user.role.upper() != "ADMIN":
+        return jsonify({"success": False, "message": "Unauthorized action"}), 403
+
+    if request.method == "POST":
+        new_data = request.get_json()
+        todo = Todos(
+            title=new_data["title"],
+            description=new_data["description"],
+            project_id=new_data["project_id"],
+        )
+
+        try:
+            db.session.add(todo)
+            db.session.commit()
+        except:
+            db.session.rollback()
+
+            flash("Failed to add the task", category="error")
+
+            return (
+                jsonify({"success": False, "message": "Failed to add the task"}),
+                500,
+            )
+        else:
+            flash(f"Your new task has been added to the list!", category="success")
+
+            return jsonify({"success": True, "data": todo.serialize()}), 201
+
     todos = db.session.execute(
         db.select(Todos)
         .join(Projects, Todos.project_id == Projects.project_id)
@@ -87,7 +227,7 @@ def get_all_todos(user_id):
         serial.update({"project_title": todo.projects.title})
         data.append(serial)
 
-    return jsonify(data)
+    return jsonify({"success": True, "data": data}), 200
 
 
 @api_bp.route(
@@ -95,7 +235,11 @@ def get_all_todos(user_id):
     methods=["GET"],
     strict_slashes=False,
 )
+@jwt_required()
 def get_todos(user_id, project_id):
+    if current_user.user_id != user_id and current_user.role.upper() != "ADMIN":
+        return jsonify({"success": False, "message": "Unauthorized action"}), 403
+
     todos = db.session.execute(
         db.select(Todos)
         .join(Projects, Todos.project_id == Projects.project_id)
@@ -108,49 +252,75 @@ def get_todos(user_id, project_id):
         serial.update({"project_title": todo.projects.title})
         data.append(serial)
 
-    return jsonify(data)
+    return jsonify({"success": True, "data": data}), 200
 
 
 @api_bp.route(
-    "/todos/<int:todo_id>",
-    methods=["GET", "POST", "PUT"],
+    "/users/<int:user_id>/todos/<int:todo_id>",
+    methods=["GET", "PUT", "DELETE"],
     strict_slashes=False,
 )
-def get_todo(todo_id):
+@jwt_required()
+def access_todo(user_id, todo_id):
+    if current_user.user_id != user_id and current_user.role.upper() != "ADMIN":
+        return jsonify({"success": False, "message": "Unauthorized action"}), 403
+
     todo = db.session.execute(
         db.select(Todos).filter(Todos.todo_id == todo_id)
     ).scalar_one()
     data = todo.serialize()
 
     if request.method == "PUT":
-        updated_data = json.loads(request.get_data(as_text=True))
-        todo.is_done = updated_data["is_done"]
+        # updated_data = json.loads(request.get_data(as_text=True))
+        try:
+            updated_data = request.get_json()
+            todo.title = updated_data["title"]
+            todo.description = updated_data["description"]
+            todo.project_id = updated_data["project_id"]
+            todo.is_done = updated_data["is_done"]
 
-        db.session.commit()
-        updated_data = todo.serialize()
+            db.session.commit()
+        except:
+            db.session.rollback()
 
-        return jsonify(updated_data), 201
+            flash("Failed to update the task", category="error")
 
-    return jsonify(data), 200
+            return (
+                jsonify({"success": False, "message": "Failed to update the task"}),
+                500,
+            )
+        else:
+            updated_data = todo.serialize()
+
+            flash(f"Your task has been updated!", category="success")
+
+            return jsonify({"success": True, "data": updated_data}), 201
+
+    elif request.method == "DELETE":
+        try:
+            db.session.delete(todo)
+            db.session.commit()
+        except:
+            db.session.rollback()
+
+            flash("Failed to delete the task", category="error")
+
+            return (
+                jsonify({"success": False, "message": "Failed to delete the task"}),
+                500,
+            )
+        else:
+            flash(f"Your task has been deleted!", category="error")
+
+            return (
+                jsonify({"success": True, "message": "Your task has been deleted!"}),
+                201,
+            )
+
+    return jsonify({"success": True, "data": data}), 200
 
 
-@api_bp.route(
-    "/todos/<int:todo_id>",
-    methods=["DELETE"],
-    strict_slashes=False,
-)
-def delete_todo(todo_id):
-    todo = db.session.execute(
-        db.select(Todos).filter(Todos.todo_id == todo_id)
-    ).scalar_one()
-    data = todo.serialize()
-
-    db.session.delete(todo)
-    db.session.commit()
-
-    flash(f"Your task has been deleted!", category="error")
-
-    return jsonify(data), 201
+##### DONES ENDPOINTS #####
 
 
 @api_bp.route(
@@ -158,7 +328,11 @@ def delete_todo(todo_id):
     methods=["GET"],
     strict_slashes=False,
 )
+@jwt_required()
 def get_all_dones(user_id):
+    if current_user.user_id != user_id and current_user.role.upper() != "ADMIN":
+        return jsonify({"success": False, "message": "Unauthorized action"}), 403
+
     dones = db.session.execute(
         db.select(Todos)
         .join(Projects, Todos.project_id == Projects.project_id)
@@ -172,7 +346,7 @@ def get_all_dones(user_id):
         serial.update({"project_title": done.projects.title})
         data.append(serial)
 
-    return jsonify(data)
+    return jsonify({"success": True, "data": data}), 200
 
 
 @api_bp.route(
@@ -180,7 +354,11 @@ def get_all_dones(user_id):
     methods=["GET"],
     strict_slashes=False,
 )
+@jwt_required()
 def get_dones(user_id, project_id):
+    if current_user.user_id != user_id and current_user.role.upper() != "ADMIN":
+        return jsonify({"success": False, "message": "Unauthorized action"}), 403
+
     dones = db.session.execute(
         db.select(Todos)
         .join(Projects, Todos.project_id == Projects.project_id)
@@ -197,4 +375,4 @@ def get_dones(user_id, project_id):
         serial.update({"project_title": done.projects.title})
         data.append(serial)
 
-    return jsonify(data)
+    return jsonify({"success": True, "data": data}), 200
